feat: model data type 변경

+++ backend_py/app/models.py
@@ -1,27 +1,27 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, ForeignKey, Boolean
+from sqlalchemy import Column, Integer, String, BigInteger, ForeignKey, Boolean
 from sqlalchemy.orm import relationship
 from app.database import Base, BaseTimeEntity
 
 class Team(Base, BaseTimeEntity):
     __tablename__ = "team"
 
-    team_id = Column(Integer, primary_key=True)
+    team_id = Column(BigInteger, primary_key=True)
     team_name = Column(String(30), nullable=False)
     song_name = Column(String(30), nullable=True)
     profile_url = Column(String(255), nullable=True)
     is_finished = Column(Boolean, nullable=False)
-    team_leader_id = Column(Integer, nullable=False)
+    team_leader_id = Column(BigInteger, nullable=False)
     scores = relationship("Score", back_populates="team")
     positions = relationship("Position", back_populates="team")
 
 class Score(Base, BaseTimeEntity):
     __tablename__ = "score"
 
-    score_id = Column(Integer, primary_key=True)
+    score_id = Column(BigInteger, primary_key=True)
     part_num = Column(Integer, nullable=False)
-    position_id = Column(Integer, ForeignKey("position.position_id"))
+    position_id = Column(BigInteger, ForeignKey("position.position_id"))
     position = relationship("Position", back_populates="score")
-    team_id = Column(Integer, ForeignKey("team.team_id"))
+    team_id = Column(BigInteger, ForeignKey("team.team_id"))
     team = relationship("Team", back_populates="scores")
     title = Column(String(30), nullable=False)
     score_url = Column(String(255), nullable=False)
@@ -31,7 +31,7 @@ class Score(Base, BaseTimeEntity):
 class Accompainment(Base, BaseTimeEntity):
     __tablename__ = "accompainment"
 
-    accompainment_id = Column(Integer, primary_key=True)
+    accompainment_id = Column(BigInteger, primary_key=True)
     score_id = Column(Integer, ForeignKey("score.score_id"))
     score = relationship("Score", back_populates="accompainment")
     accompainment_url = Column(String(255), nullable=False)
@@ -39,18 +39,18 @@ class Accompainment(Base, BaseTimeEntity):
 class Position(Base, BaseTimeEntity):
     __tablename__ = "position"
 
-    position_id = Column(Integer, primary_key=True)
+    position_id = Column(BigInteger, primary_key=True)
     position_name = Column(String(30), nullable=False)
-    team_id = Column(Integer, ForeignKey("team.team_id"))
+    team_id = Column(BigInteger, ForeignKey("team.team_id"))
     team = relationship("Team", back_populates="positions")
-    color_id = Column(Integer, ForeignKey("color.color_id"))
+    color_id = Column(BigInteger, ForeignKey("color.color_id"))
     color = relationship("Color", back_populates="position")
     score = relationship("Score", back_populates="position")
 
 class Color(Base, BaseTimeEntity):
     __tablename__ = "color"
 
-    color_id = Column(Integer, primary_key=True)
+    color_id = Column(BigInteger, primary_key=True)
     color_name = Column(String(255), nullable=False)
     color_code = Column(String(255), nullable=False)
     position = relationship("Position", back_populates="color", uselist=False)
